Going to swe layer now changes windows lang too
_LAYER6 is my swedish layer with åäö on and nothing else special.
Need to change windows to swedish locale however and now this is done
when changing to that layer and back.

It can get out of sync if you run win+space manually. But I prefer that
to setting custom keybinds to switch to a specific language in windows.
Its just a win+space toggle away in that case so thats fine.

The main reason for this layer is that I want to use backspace in the
top right corner for normal english use. But when writing swedish i need
that key for å. That is a seperate issue that I want to fix as well,
using backspace when writing swedish..